ProjectCommon made dynamic MyCommon is ancestor for classes using MyCMS, i.e. MyController and (the new) MyAdmin
